Implement VAR for C21_WMDE_Test_12
tests white collage banner leading to AB test

https://phabricator.wikimedia.org/T295869
