+++ desktop/components/Banner.jsx
@@ -1,32 +1,19 @@
-import { Component, createRef, h } from 'preact';
+import { Component, h, createRef } from 'preact';
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
 import BannerTransition from '../../shared/components/BannerTransition';
-import ProgressBar from './ui/ProgressBar';
-import CloseIcon from './ui/CloseIcon';
+import Infobox from './ui/Infobox';
 import FundsDistributionInfo from '../../shared/components/ui/use_of_funds/FundsDistributionInfo';
 import FundsModal from '../../shared/components/ui/use_of_funds/FundsModal';
 import TranslationContext from '../../shared/components/TranslationContext';
-import debounce from '../../shared/debounce';
-import ChevronLeftIcon from './ui/ChevronLeftIcon';
-import SlideState from '../../shared/slide_state';
 import { BannerType } from '../BannerType';
-import ContentSlideshow from './ui/ContentSlideshow';
-import ContentForm from './ui/ContentForm';
 import createDynamicCampaignText from '../create_dynamic_campaign_text';
-
-const SLIDESHOW_START_DELAY = 2000;
-const SLIDESHOW_SLIDE_INTERVAL = 10000;
+import ProgressBar from '../../shared/components/ui/ProgressBar';
 
 const BannerVisibilityState = Object.freeze( {
 	PENDING: Symbol( 'pending' ),
 	VISIBLE: Symbol( 'visible' ),
 	CLOSED: Symbol( 'closed' )
-} );
-
-const BannerContentState = Object.freeze( {
-	SLIDES: Symbol( 'slides' ),
-	FORM: Symbol( 'form' )
 } );
 
 export class Banner extends Component {
@@ -40,25 +27,18 @@ export class Banner extends Component {
 		registerDisplayBanner: PropTypes.func.isRequired
 	}
 
-	bannerRef = createRef();
-	contentSlideshowRef = createRef();
-	contentFormRef = createRef();
+	ref = createRef();
 
 	constructor( props ) {
 		super( props );
 		this.state = {
 			bannerVisibilityState: BannerVisibilityState.PENDING,
-			bannerContentState: BannerContentState.SLIDES,
 			isFundsModalVisible: false,
-			contentSize: 'auto',
 
 			// trigger for banner resize events
 			formInteractionSwitcher: false
 		};
 		this.slideInBanner = () => {};
-		this.startSliderAutoplay = () => {};
-		this.stopSliderAutoplay = () => {};
-		this.slideState = new SlideState();
 		this.dynamicCampaignText = createDynamicCampaignText(
 			props.campaignParameters,
 			props.campaignProjection,
@@ -75,62 +55,30 @@ export class Banner extends Component {
 				this.slideInBanner();
 			}
 		);
-		this.props.registerResizeBanner( debounce( this.onPageResize.bind( this ), 200 ) );
-		this.setContentSize();
+		this.props.registerResizeBanner( this.adjustSurroundingSpace.bind( this ) );
 	}
 
-	trackBannerEvent( eventName ) {
-		this.props.trackingData.tracker.trackBannerEvent(
-			eventName,
-			this.slideState.slidesShown,
-			this.slideState.currentSlide + 1,
-			this.props.trackingData.bannerClickTrackRatio
-		);
-	}
-
-	onPageResize() {
-		if ( this.state.bannerVisibilityState !== BannerVisibilityState.VISIBLE ) {
-			return;
-		}
-
-		this.setContentSize();
-		this.addBannerSpace();
-	}
-
-	addBannerSpace() {
+	adjustSurroundingSpace() {
 		const bannerElement = document.querySelector( '.wmde-banner .banner-position' );
 		this.props.skinAdjuster.addSpaceInstantly( bannerElement.offsetHeight );
-	}
-
-	setContentSize() {
-		let height = this.contentSlideshowRef.current.clientHeight;
-		if ( this.state.bannerContentState === BannerContentState.FORM ) {
-			height = this.contentFormRef.current.clientHeight;
-		}
-		this.setState( { contentSize: `${height}px` } );
 	}
 
 	// eslint-disable-next-line no-unused-vars
 	componentDidUpdate( previousProps, previousState, snapshot ) {
 		if ( previousState.formInteractionSwitcher !== this.state.formInteractionSwitcher ) {
-			this.onPageResize();
+			this.adjustSurroundingSpace();
 		}
 	}
 
 	onFinishedTransitioning = () => {
 		this.props.onFinishedTransitioning();
 		this.startProgressbar();
-		setTimeout( this.startSliderAutoplay, SLIDESHOW_START_DELAY );
-		this.onPageResize();
 	}
 
 	closeBanner = e => {
 		e.preventDefault();
 		this.setState( { bannerVisibilityState: BannerVisibilityState.CLOSED } );
-		this.props.onClose(
-			this.slideState.slidesShown,
-			this.slideState.currentSlide
-		);
+		this.props.onClose();
 	};
 
 	registerBannerTransition = ( cb ) => {
@@ -141,45 +89,25 @@ export class Banner extends Component {
 		this.startProgressbar = startPb;
 	};
 
-	registerAutoplayCallbacks = ( onStartAutoplay, onStopAutoplay ) => {
-		this.startSliderAutoplay = onStartAutoplay;
-		this.stopSliderAutoplay = onStopAutoplay;
-	};
-
 	toggleFundsModal = () => {
 		if ( !this.state.isFundsModalVisible ) {
-			this.trackBannerEvent( 'application-of-funds-shown' );
+			this.props.trackingData.tracker.trackBannerEvent( 'application-of-funds-shown', 0, 0, this.props.trackingData.bannerClickTrackRatio );
 		}
 		this.setState( { isFundsModalVisible: !this.state.isFundsModalVisible } );
 	};
 
 	fundsModalDonate = () => {
-		this.trackBannerEvent( 'funds-modal-donate-clicked' );
-		this.setState( { isFundsModalVisible: false, bannerContentState: BannerContentState.FORM }, this.setContentSize );
+		this.props.trackingData.tracker.trackBannerEvent( 'funds-modal-donate-clicked', 0, 0, this.props.trackingData.bannerClickTrackRatio );
+		this.setState( { isFundsModalVisible: false } );
 	};
 
 	onFormInteraction = () => {
 		this.setState( { showLanguageWarning: true, formInteractionSwitcher: !this.state.formInteractionSwitcher } );
 	}
 
-	showDonationForm = e => {
-		e.preventDefault();
-		this.trackBannerEvent( 'desktop-banner-expanded' );
-		this.setState( { bannerContentState: BannerContentState.FORM }, this.setContentSize );
-		this.stopSliderAutoplay();
-	}
-
-	showSlides = e => {
-		e.preventDefault();
-		this.setState( { bannerContentState: BannerContentState.SLIDES }, this.setContentSize );
-	}
-
-	onSlideChange = ( index ) => {
-		this.slideState.onSlideChange( index );
-	}
-
 	// eslint-disable-next-line no-unused-vars
 	render( props, state, context ) {
+		const DonationForm = props.donationForm;
 		const campaignProjection = props.campaignProjection;
 		const Footer = props.footer;
 
@@ -188,12 +116,10 @@ export class Banner extends Component {
 				'wmde-banner': true,
 				'wmde-banner--hidden': state.bannerVisibilityState === BannerVisibilityState.CLOSED,
 				'wmde-banner--visible': state.bannerVisibilityState === BannerVisibilityState.VISIBLE,
-				'wmde-banner--content-slides': state.bannerContentState === BannerContentState.SLIDES,
-				'wmde-banner--content-form': state.bannerContentState === BannerContentState.FORM,
 				'wmde-banner--ctrl': props.bannerType === BannerType.CTRL,
 				'wmde-banner--var': props.bannerType === BannerType.VAR
 			} ) }
-			ref={this.bannerRef}>
+			ref={this.ref}>
 			<BannerTransition
 				fixed={ true }
 				registerDisplayBanner={ this.registerBannerTransition }
@@ -203,68 +129,49 @@ export class Banner extends Component {
 			>
 				<TranslationContext.Provider value={props.translations}>
 					<div className="banner__wrapper">
-						<div className="banner__inner">
-							<div className="banner__navigation">
-								<div className="banner__back">
-									<a className="back__link" onClick={ this.showSlides }>
-										<ChevronLeftIcon/> { props.translations[ 'back-button' ] }
-									</a>
-								</div>
-								<div className="banner__close">
-									<a className="close-link" onClick={ this.closeBanner }>
-										{ props.translations[ 'close-button' ] } <CloseIcon/>
-									</a>
+						<div className="close">
+							<a className="close__link" onClick={this.closeBanner}>&#x2715;</a>
+						</div>
+						<div className="banner__content">
+							<div className="banner__infobox">
+								<div className="infobox-bubble">
+									<Infobox
+										formatters={props.formatters}
+										campaignParameters={props.campaignParameters}
+										campaignProjection={props.campaignProjection}
+										bannerText={props.bannerText}
+										dynamicCampaignText={ this.dynamicCampaignText }
+										propsForText={ {
+											overallImpressionCount: props.impressionCounts.getOverallCount(),
+											millionImpressionsPerDay: props.campaignParameters.millionImpressionsPerDay
+										} }/>
 								</div>
 							</div>
-							<div className="banner__logo">
-								<img src="https://upload.wikimedia.org/wikipedia/commons/1/10/Wikipedia-logo-compressed.png" alt="Wikipedia Logo"/>
-							</div>
-							<div className="banner__inner-content-size-fitter" style={ `height:${ state.contentSize }` } onTransitionEnd={ () => {
-								this.onPageResize();
-							} }>
-								<ContentSlideshow
-									ref={ this.contentSlideshowRef }
-									campaignParameters={ props.campaignParameters }
-									campaignProjection={ props.campaignProjection }
-									formatters={ props.formatters }
-									slideshowSlideInterval={ SLIDESHOW_SLIDE_INTERVAL }
-									toggleFundsModal={ this.toggleFundsModal }
-									showDonationForm={ this.showDonationForm }
-									registerAutoplayCallbacks={ this.registerAutoplayCallbacks }
-									onSlideChange={ this.onSlideChange }
-									dynamicCampaignText={ this.dynamicCampaignText }
-								/>
-								<ContentForm
-									ref={ this.contentFormRef }
-									formatters={ props.formatters }
-									campaignParameters={ props.campaignParameters }
-									campaignProjection={ props.campaignProjection }
-									bannerText={ props.bannerText }
-									propsForText={ {
-										dynamicCampaignText: this.dynamicCampaignText,
-										toggleFundsModal: this.toggleFundsModal
-									} }
-									donationForm={ props.donationForm }
-									formItems={ props.formItems }
-									bannerName={ props.bannerName }
-									campaignName={ props.campaignName }
-									impressionCounts={ props.impressionCounts }
-									onFormInteraction={ this.onFormInteraction }
-									onSubmit={ props.onSubmit }
+							<div className="banner__form">
+								<DonationForm
+									formItems={props.formItems}
+									bannerName={props.bannerName}
+									campaignName={props.campaignName}
+									formatters={props.formatters}
+									impressionCounts={props.impressionCounts}
+									onFormInteraction={this.onFormInteraction}
+									onSubmit={props.onSubmit}
+									customAmountPlaceholder={ props.translations[ 'custom-amount-placeholder' ] }
 									buttonText={ props.buttonText }
 									errorPosition={ props.errorPosition }
+									bannerType={ props.bannerType }
+									showCookieBanner={ props.showCookieBanner }
 								/>
 							</div>
-							<Footer/>
 						</div>
+						<Footer showFundsModal={ this.toggleFundsModal }/>
 						<ProgressBar
 							formatters={props.formatters}
 							daysLeft={campaignProjection.getRemainingDays()}
 							donationAmount={campaignProjection.getProjectedDonationSum()}
 							goalDonationSum={campaignProjection.goalDonationSum}
 							missingAmount={campaignProjection.getProjectedRemainingDonationSum()}
-							setStartAnimation={this.registerStartProgressbar}
-						/>
+							setStartAnimation={this.registerStartProgressbar}/>
 					</div>
 				</TranslationContext.Provider>
 			</BannerTransition>
