+++ desktop/components/BannerText.jsx
@@ -1,10 +1,7 @@
 import { h } from 'preact';
-import { useContext } from 'preact/hooks';
-import TranslationContext from '../../shared/components/TranslationContext';
 import InfoIcon from './ui/InfoIcon';
 
 export default function BannerText( props ) {
-	const Translations = useContext( TranslationContext );
 	const { currentDayName, campaignDaySentence, visitorsVsDonorsSentence, overallImpressionCount } = props.dynamicCampaignText;
 
 	return <div className="banner-text">
@@ -12,18 +9,17 @@ export default function BannerText( props ) {
 			<strong><InfoIcon/> An alle, die Wikipedia in Deutschland nutzen </strong>
 		</p>
 		<p>
-			Vielleicht kommen wir gerade ungelegen, aber dennoch: Klicken Sie jetzt bitte nicht weg! Dies ist der { overallImpressionCount }.
-			Spendenaufruf, den wir Ihnen zeigen. Am heutigen { currentDayName } bitten wir Sie bescheiden, die Unabhängigkeit von Wikipedia zu
-			sichern. { campaignDaySentence } Wikipedia wird durch Spenden von durchschnittlich 21,60&nbsp;€ finanziert. Insgesamt spenden 99% nichts
-			- sie übergehen diesen Aufruf. Sollten Sie zu dem kleinen Kreis gehören, die bereits gespendet haben, danken wir Ihnen sehr
-			herzlich. { visitorsVsDonorsSentence } Die meisten Menschen spenden, weil sie Wikipedia nützlich finden. Schon mit einer Spende
-			von 5&nbsp;€ kann Wikipedia sich auch in Zukunft erfolgreich entwickeln. Hat Wikipedia Ihnen in diesem Jahr Wissen im
-			Wert einer Tasse Kaffee geschenkt? Dann nehmen Sie sich doch bitte eine Minute Zeit und geben Sie etwas zurück. Vielen Dank.
-		</p>
-		<p>
-			<a onClick={ props.toggleFundsModal }>
-				{ Translations[ 'use-of-funds-link' ] }
-			</a>
+			Vielleicht kommen wir gerade ungelegen, aber dennoch: Klicken Sie jetzt bitte nicht weg!
+			Dies ist der { overallImpressionCount }. Spendenaufruf, den wir Ihnen zeigen. Am
+			heutigen { currentDayName } bitten wir Sie bescheiden, die Unabhängigkeit von Wikipedia
+			zu sichern. { campaignDaySentence } Insgesamt
+			spenden 99% nichts – sie übergehen diesen Aufruf. Sollten Sie zu dem kleinen Kreis gehören, die
+			bereits gespendet haben, danken wir Ihnen sehr herzlich. Wikipedia wird durch Spenden von
+			durchschnittlich 21,60&nbsp;€ finanziert. Doch schon mit einer Spende von 5&nbsp;€ kann Wikipedia sich auch
+			in Zukunft erfolgreich entwickeln. { visitorsVsDonorsSentence } Die meisten Menschen spenden, weil
+			sie Wikipedia nützlich finden. Hat Wikipedia Ihnen in diesem Jahr Wissen im Wert einer Tasse
+			Kaffee geschenkt? Dann nehmen Sie sich doch bitte eine Minute Zeit und geben Sie
+			etwas zurück. Vielen Dank
 		</p>
 	</div>;
 }
