--- desktop/components/ui/ChevronDownIcon.jsx
@@ -1,8 +1,0 @@
-import { h } from 'preact';
-
-export default function ChevronDownIcon() {
-	return <svg width="22" height="14" viewBox="0 0 22 14" fill="none" xmlns="http://www.w3.org/2000/svg">
-		{/* eslint-disable-next-line max-len */}
-		<path d="M10.9299 13.8595C10.2619 13.8595 9.62189 13.5845 9.16289 13.0985L0.662885 4.0985C-0.258115 3.1225 -0.214115 1.5855 0.761885 0.663497C1.73789 -0.257503 3.27589 -0.213503 4.19689 0.760497L10.9299 7.8905L17.6629 0.760497C18.5849 -0.213503 20.1219 -0.258503 21.0979 0.663497C22.0739 1.5855 22.1179 3.1235 21.1969 4.0985L12.6969 13.0985C12.2379 13.5845 11.5979 13.8595 10.9299 13.8595Z" fill="white"/>
-	</svg>;
-}
