--- desktop/components/ui/ChevronLeftIcon.jsx
@@ -1,8 +1,0 @@
-import { h } from 'preact';
-
-export default function ChevronLeftIcon() {
-	return <svg width="15" height="23" viewBox="0 0 15 23" fill="none" xmlns="http://www.w3.org/2000/svg">
-		{/* eslint-disable-next-line max-len */}
-		<path d="M11.9994 22.9996C12.7954 22.9996 13.5904 22.6846 14.1814 22.0596C15.3184 20.8546 15.2644 18.9566 14.0604 17.8186L7.3694 11.4996L14.0604 5.1806C15.2644 4.0426 15.3184 2.1446 14.1814 0.939601C13.0424 -0.264399 11.1454 -0.318399 9.9394 0.818601L0.939399 9.3186C0.339398 9.8856 -0.000600815 10.6746 -0.000600815 11.4996C-0.000600815 12.3246 0.339398 13.1136 0.939399 13.6806L9.9394 22.1806C10.5204 22.7286 11.2604 22.9996 11.9994 22.9996Z" fill="#3366CC"/>
-	</svg>;
-}
