+++ desktop/components/ui/form/DonationForm.jsx
@@ -1,10 +1,9 @@
 import { h } from 'preact';
-import { useContext, useState } from 'preact/hooks';
-import classNames from 'classnames';
+import { useContext } from 'preact/hooks';
 
 import TranslationContext from '../../../../shared/components/TranslationContext';
-import { SelectGroup } from './SelectGroup';
-import SelectCustomAmount from './SelectCustomAmount';
+import { SelectGroup } from '../../../../shared/components/ui/form/SelectGroup';
+import SelectCustomAmount from '../../../../shared/components/ui/form/SelectCustomAmount';
 import SmsBox from '../../../../shared/components/ui/form/SmsBox';
 
 import { isValid, isValidOrUnset } from '../../../../shared/components/ui/form/hooks/validation_states';
@@ -12,9 +11,7 @@ import useAmountWithCustom from '../../../../shared/components/ui/form/hooks/use
 import useInterval from '../../../../shared/components/ui/form/hooks/use_interval';
 import usePaymentMethod from '../../../../shared/components/ui/form/hooks/use_payment_method';
 import { amountMessage, validateRequired } from '../../../../shared/components/ui/form/utils';
-import { Intervals, PaymentMethods } from '../../../../shared/components/ui/form/FormItemsBuilder';
 import SubmitValues from '../../../../shared/components/ui/form/SubmitValues';
-import ChevronRightIcon from '../ChevronRightIcon';
 import useFormAction from '../../../../shared/components/ui/form/hooks/use_form_action';
 
 export default function DonationForm( props ) {
@@ -25,13 +22,9 @@ export default function DonationForm( props ) {
 		{ numericAmount, amountValidity, selectedAmount, customAmount },
 		{ selectAmount, updateCustomAmount, validateCustomAmount, setAmountValidity }
 	] = useAmountWithCustom( null, props.formatters.customAmountInputFormatter );
-	const [ disabledIntervals, setDisabledIntervals ] = useState( [] );
-	const [ disabledPaymentMethods, setDisabledPaymentMethods ] = useState( [] );
+	const disabledIntervals = [];
+	const disabledPaymentMethods = [];
 	const [ formAction ] = useFormAction( props, { ocn: props.showCookieBanner } );
-
-	const isFormValid = () => {
-		return isValid( intervalValidity ) && isValid( amountValidity ) && isValid( paymentMethodValidity );
-	};
 
 	const validate = e => {
 		if ( [
@@ -44,49 +37,25 @@ export default function DonationForm( props ) {
 		}
 		e.preventDefault();
 	};
-
-	const onChangeInterval = e => {
-		setInterval( e.target.value );
-		if ( e.target.value !== Intervals.ONCE.value ) {
-			setDisabledPaymentMethods( [ PaymentMethods.SOFORT.value ] );
-		} else {
-			setDisabledPaymentMethods( [] );
-		}
-	};
-
-	const onChangePaymentMethod = e => {
-		setPaymentMethod( e.target.value );
-		if ( e.target.value === PaymentMethods.SOFORT.value ) {
-			setDisabledIntervals(
-				// Exclude all intervals except "once"
-				props.formItems.intervals
-					.map( intervalItem => intervalItem.value )
-					.filter( interval => interval !== Intervals.ONCE.value )
-			);
-		} else {
-			setDisabledIntervals( [] );
-		}
-	};
+	const onFormInteraction = this.props.onFormInteraction ? e => this.props.onFormInteraction( e ) : () => {};
 
 	return <div className="form">
-		<form method="post" name="donationForm" className="form__element" action={ formAction }>
+		<form method="post" name="donationForm" className="form__element" onClick={ onFormInteraction } action={ formAction }>
 
-			<fieldset className="form__section">
-				<legend className="form__section-head">{ Translations[ 'intervals-header' ]}</legend>
+			<div className="form-field-group">
 				<SelectGroup
 					fieldname="select-interval"
 					selectionItems={ props.formItems.intervals }
 					isValid={ isValidOrUnset( intervalValidity ) }
 					errorMessage={ Translations[ 'no-interval-message' ] }
 					currentValue={ paymentInterval }
-					onSelected={ onChangeInterval }
+					onSelected={ e => setInterval( e.target.value ) }
 					disabledOptions={ disabledIntervals }
 					errorPosition={ props.errorPosition }
 				/>
-			</fieldset>
+			</div>
 
-			<fieldset className="form__section">
-				<legend className="form__section-head">{ Translations[ 'amounts-header' ]}</legend>
+			<div className={ 'form-field-group' }>
 				<SelectGroup
 					fieldname="select-amount"
 					selectionItems={props.formItems.amounts}
@@ -110,27 +79,26 @@ export default function DonationForm( props ) {
 						}
 					/>
 				</SelectGroup>
-			</fieldset>
+			</div>
 
-			<fieldset className="form__section">
-				<legend className="form__section-head">{ Translations[ 'payments-header' ] }</legend>
+			<div className="form-field-group">
 				<SelectGroup
 					fieldname="select-payment-method"
 					selectionItems={ props.formItems.paymentMethods }
 					isValid={ isValidOrUnset( paymentMethodValidity )}
 					errorMessage={ Translations[ 'no-payment-type-message' ] }
 					currentValue={ paymentMethod }
-					onSelected={ onChangePaymentMethod }
+					onSelected={ e => setPaymentMethod( e.target.value ) }
 					disabledOptions={ disabledPaymentMethods }
 					errorPosition={ props.errorPosition }
 				>
 					<SmsBox/>
 				</SelectGroup>
-			</fieldset>
+			</div>
 
 			<div className="submit-section button-group">
-				<button className={ classNames( 'button-group__button', { 'is-valid': isFormValid() } ) } onClick={ validate }>
-					<span className="button-group__label">{ Translations[ 'submit-label' ] } <ChevronRightIcon/></span>
+				<button className="button-group__button" onClick={ validate }>
+					<span className="button-group__label">{ Translations[ 'submit-label' ] }</span>
 				</button>
 			</div>
