+++ desktop/styles/SelectGroup.pcss
