+++ desktop/styles/Slider.pcss
@@ -11,7 +11,6 @@
 		-khtml-user-select: none;
 		touch-action: pan-y;
 		-webkit-tap-highlight-color: transparent;
-		align-items: center;
 	}
 
 	.keen-slider,
@@ -24,8 +23,6 @@
 	.keen-slider__slide {
 		width: 100%;
 		min-height: 100%;
-		min-width: calc( 100% + 0px );
-		max-width: calc( 100% + 0px );
 	}
 
 	.keen-slider__slide-content {
@@ -35,23 +32,6 @@
 		justify-content: center;
 		align-items: center;
 		cursor: grab;
-		font-family: $banner-serif;
-		font-size: 17px;
-		text-align: center;
-		padding: 0 26px 16px;
-
-		@media ( min-width: $breakpoint_m ) {
-			padding: 16px 30px 10px 36px;
-		}
-
-		.headline {
-			font-size: 22px;
-			line-height: 41px;
-
-			@media ( min-width: $breakpoint_xs ) {
-				font-size: 26px;
-			}
-		}
 	}
 
 	.keen-slider[ data-keen-slider-v ] {
@@ -66,94 +46,69 @@
 		pointer-events: none;
 	}
 
-	.slider-navigation-previous,
-	.slider-navigation-next {
-		position: absolute;
-		z-index: 30;
-		width: 52px;
-		height: 52px;
-		line-height: 52px;
-		top: 50%;
-		margin-top: -26px;
-		border: 0;
-		background: $color-white;
-		border-radius: 50%;
-		text-align: center;
-		padding: 5px 0 0;
-	}
-
-	.slider-navigation-previous {
-		left: -26px;
-		padding-right: 10px;
-		box-shadow: -3px 0 3px rgba( 0, 0, 0, 0.1 );
-		transition: left 500ms $banner-easing;
-
-		&:hover,
-		&:focus {
-			left: -30px;
-		}
-	}
-
-	.slider-navigation-next {
-		right: -26px;
-		padding-left: 10px;
-		box-shadow: 3px 0 3px rgba( 0, 0, 0, 0.1 );
-		transition: right 500ms $banner-easing;
-
-		&:hover,
-		&:focus {
-			right: -30px;
-		}
-	}
-
-	.slideshow-application-of-funds-link {
-		position: absolute;
-		display: inline-block;
-		bottom: 14px;
-		left: 20px;
-		font-size: 12px;
-
-		@media ( min-width: $breakpoint_xs ) {
-			font-size: 15px;
-		}
+	.slider-container,
+	.navigation-wrapper {
+		height: 100%;
+		width: 100%;
 	}
 
 	.slider-container {
+		position: absolute;
+		padding-bottom: 20px;
+	}
+
+	.navigation-wrapper {
 		position: relative;
-		padding-bottom: 28px;
-		@media ( min-width: $breakpoint_xs ) {
-			padding-bottom: 38px;
+		padding: 0 30px;
+	}
+
+	.slider-navigation-previous,
+	.slider-navigation-next {
+		position: absolute;
+		height: 30px;
+		width: 30px;
+		line-height: 30px;
+		top: 50%;
+		margin-top: -15px;
+
+		&:hover,
+		&:focus {
+			svg path {
+				fill: $color-gray-medium;
+			}
 		}
+	}
+
+	.slider-navigation-previous {
+		left: 0;
+		text-align: left;
+	}
+
+	.slider-navigation-next {
+		right: 0;
+		text-align: right;
 	}
 
 	.pagination {
 		position: absolute;
 		bottom: 0;
 		width: 100%;
-		height: 38px;
-		margin: 0;
-		list-style: none;
 		text-align: center;
-		line-height: 1;
-
 		&-dot {
-			background: $color-white;
+			background: $color-gray-light;
 			display: inline-block;
-			width: 12px;
-			height: 12px;
+			width: 16px;
+			height: 16px;
 			padding: 0;
 			margin: 0 4px;
-			border: 2px solid $color-gray;
 			border-radius: 50%;
 			cursor: pointer;
-
-			@media ( min-width: $breakpoint_xs ) {
-				width: 20px;
-				height: 20px;
-			}
+			border: 4px solid $color-white;
+			box-sizing: border-box;
 
 			&.is-active {
-				background: $color-gray;
+				background: $color-white;
+				border: 3px solid $color-primary;
 			}
 		}
 	}
