+++ desktop/styles/styles.pcss
@@ -3,14 +3,14 @@
 
 @import '../../shared/components/BannerTransition.pcss';
 
-@import 'Banner.pcss';
-@import 'BannerNavigation.pcss';
-@import 'Slider.pcss';
-@import 'ProgressBar.pcss';
-@import 'BannerText.pcss';
+@import './Banner.pcss';
 @import 'DonationForm.pcss';
-@import 'Footer.pcss';
-
+@import './SelectGroup.pcss';
+@import './Footer.pcss';
+@import './Infobox.pcss';
+@import 'BannerText.pcss';
+@import './ProgressBar.pcss';
+@import 'Slider.pcss';
 @import '../../shared/components/ui/use_of_funds/FundsModal.pcss';
 @import '../../shared/components/ui/use_of_funds/FundsContent.pcss';
 @import '../../shared/components/ui/use_of_funds/FundsDistributionInfo.pcss';
