--- desktop/styles/styles_var.pcss
@@ -1,17 +1,0 @@
-@import '../../shared/components/ui/use_of_funds/variables.pcss';
-@import 'variables.pcss';
-
-@import '../../shared/components/BannerTransition.pcss';
-
-@import 'Banner.pcss';
-@import 'BannerNavigation_var.pcss';
-@import 'Slider.pcss';
-@import 'ProgressBar.pcss';
-@import 'BannerText.pcss';
-@import 'DonationForm.pcss';
-@import 'Footer.pcss';
-
-@import '../../shared/components/ui/use_of_funds/FundsModal.pcss';
-@import '../../shared/components/ui/use_of_funds/FundsContent.pcss';
-@import '../../shared/components/ui/use_of_funds/FundsDistributionInfo.pcss';
-@import '../../shared/components/ui/use_of_funds/CompanyBudgets.pcss';
