--- desktop/styles_var/Footer.pcss
@@ -1,100 +1,0 @@
-.wmde-banner {
-	.footer {
-		display: flex;
-		flex-wrap: wrap;
-		flex-direction: row;
-		justify-content: flex-start;
-		font-size: 12px;
-		line-height: 20px;
-		color: $color-black;
-		padding: 0 10px;
-		margin-top: -10px;
-
-		@media ( min-width: $breakpoint_l ) {
-			padding: 0 24px;
-			margin-top: 0;
-		}
-
-		&__bank,
-		&__usage {
-			padding: 10px 0;
-		}
-
-		&__bank {
-			flex: 1 1 auto;
-			display: flex;
-			flex-wrap: wrap;
-
-			@media ( min-width: $breakpoint_l ) {
-				border-right: 1px solid $color-secondary-light;
-				display: block;
-				flex-wrap: nowrap;
-				margin-right: 30px;
-			}
-			&-item {
-				white-space: nowrap;
-				padding-right: 2em;
-
-				&.account .selection-input-text,
-				&.account .selection-input-input {
-					text-transform: none;
-				}
-			}
-		}
-
-		&__usage {
-			flex-grow: 1;
-			text-align: right;
-			@media ( min-width: $breakpoint_l ) {
-				flex: 0 0 $form-width;
-				text-align: center;
-				width: $form-width;
-			}
-		}
-	}
-
-	.selection-input {
-		display: inline-block;
-		position: relative;
-		margin-left: 5px;
-
-		&-text,
-		&-input {
-			font-weight: bold;
-			font-size: 12px;
-		}
-		&-input {
-			font-family: $banner-main-font-family;
-			position: absolute;
-			top: 0;
-			left: 0;
-			height: 100%;
-			width: 100%;
-			padding: 0;
-			border: 0;
-			outline: 0;
-			background: transparent;
-			color: transparent;
-			z-index: 2;
-		}
-
-		&.focused {
-			.selection-input-text {
-				color: transparent;
-			}
-			.selection-input-input {
-				color: $color-black;
-
-				&::selection {
-					background: $color-primary;
-					color: $color-white;
-				}
-			}
-		}
-	}
-
-	.application-of-funds-link {
-		color: $color-black;
-		cursor: pointer;
-	}
-}
