--- desktop/styles_var/Slider.pcss
@@ -1,115 +1,0 @@
-.wmde-banner {
-	.keen-slider {
-		display: flex;
-		height: 100%;
-		width: 100%;
-		-webkit-user-select: none;
-		-moz-user-select: none;
-		-ms-user-select: none;
-		user-select: none;
-		-webkit-touch-callout: none;
-		-khtml-user-select: none;
-		touch-action: pan-y;
-		-webkit-tap-highlight-color: transparent;
-	}
-
-	.keen-slider,
-	.keen-slider__slide {
-		overflow: hidden;
-		position: relative;
-		height: 100%;
-	}
-
-	.keen-slider__slide {
-		width: 100%;
-		min-height: 100%;
-	}
-
-	.keen-slider__slide-content {
-		position: relative;
-		height: 100%;
-		display: flex;
-		justify-content: center;
-		align-items: center;
-		cursor: grab;
-	}
-
-	.keen-slider[ data-keen-slider-v ] {
-		flex-wrap: wrap;
-	}
-
-	.keen-slider[ data-keen-slider-v ] .keen-slider__slide {
-		width: 100%;
-	}
-
-	.keen-slider[ data-keen-slider-moves ] * {
-		pointer-events: none;
-	}
-
-	.slider-container,
-	.navigation-wrapper {
-		height: 100%;
-		width: 100%;
-	}
-
-	.slider-container {
-		position: absolute;
-		padding-bottom: 20px;
-	}
-
-	.navigation-wrapper {
-		position: relative;
-		padding: 0 30px;
-	}
-
-	.slider-navigation-previous,
-	.slider-navigation-next {
-		position: absolute;
-		height: 30px;
-		width: 30px;
-		line-height: 30px;
-		top: 50%;
-		margin-top: -15px;
-
-		&:hover,
-		&:focus {
-			svg path {
-				fill: $color-gray-medium;
-			}
-		}
-	}
-
-	.slider-navigation-previous {
-		left: 0;
-		text-align: left;
-	}
-
-	.slider-navigation-next {
-		right: 0;
-		text-align: right;
-	}
-
-	.pagination {
-		position: absolute;
-		bottom: 0;
-		width: 100%;
-		text-align: center;
-		&-dot {
-			background: $color-gray-light;
-			display: inline-block;
-			width: 16px;
-			height: 16px;
-			padding: 0;
-			margin: 0 4px;
-			border-radius: 50%;
-			cursor: pointer;
-			border: 4px solid $color-white;
-			box-sizing: border-box;
-
-			&.is-active {
-				background: $color-white;
-				border: 3px solid $color-primary;
-			}
-		}
-	}
-}
